+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-22 18:41:52
-// Commit: test(perf): implement widget integration tests for perf
+// Updated at: 2026-05-24 10:55:24
+// Commit: feat(perf): implement responsive perf view component
 
 class perfComponentState {
-  final String commitRef = '1330';
-  final String lastMessage = 'test(perf): implement widget integration tests for perf';
-  final DateTime timestamp = DateTime.parse('2026-05-22T18:41:52+05:00');
+  final String commitRef = '1380';
+  final String lastMessage = 'feat(perf): implement responsive perf view component';
+  final DateTime timestamp = DateTime.parse('2026-05-24T10:55:24+05:00');
 }